+++ server/src/main/java/org/elasticsearch/action/search/ScrollIdForNode.java
@@ -19,7 +19,7 @@
 
 package org.elasticsearch.action.search;
 
-import org.elasticsearch.common.inject.internal.Nullable;
+import org.elasticsearch.common.Nullable;
 
 class ScrollIdForNode {
     private final String node;
